lkjlkj
 enter the commit message for your changes. Lines starting

+++ src/pHazardPath/HazardPath.h
@@ -23,7 +23,7 @@ class HazardPath : public CMOOSApp
   void RegisterVariables();
   void genLawnMower(int, int); 
   std::string intToString(int); 
-  std::vector<std::vector<std::vector<double> > > genTracklines(int swath);
+  std::vector<std::vector<std::vector<int> > > genTracklines(int swath);
 
 protected:
   int m_offset;
